fixed untracked precision loss risk on LP64 implementations
printf()'s "%*i", "%*u" and related specifiers need to correspond the asterisk to a minimum number of digits to use leading-zero-padding to always display, and the native type of this number to sprintf() and friends is (int).  (size_t) can, on many implementations, be larger in storage than (int), not counting the fact that unsigned integers are open to storing one of twice as many possible whole numbers as signed integers are, creating a risk of lost precision if left unchecked.